spikes: architecture digram looks nicer
This fixes the rounded rectangles not looking correctly and shows as a
proof of concept what is needed for more complicated shapes. In this
example we implement the two rounded corners of a rectangle as a svg
`path` which is abstracted away as a function. Further steps could be
made to make this a d3 shape, however is not needed at this stage.